+++ src/biquadcombo.rs
@@ -137,6 +137,35 @@ impl BiquadCombo {
         filters
     }
 
+    fn make_graphic(
+        samplerate: usize,
+        freq_min: f32,
+        freq_max: f32,
+        gains: &[f32],
+    ) -> Vec<biquad::Biquad> {
+        let bands = gains.len();
+        let mut filters = Vec::with_capacity(bands);
+
+        let f_min_log = freq_min.log2();
+        let f_max_log = freq_max.log2();
+        let bw = (f_max_log - f_min_log) / bands as f32;
+        for (band, gain) in gains.iter().enumerate() {
+            if gain.abs() > 0.001 {
+                let freq_log = f_min_log + (band as f32 + 0.5) * bw;
+                let freq = 2.0_f32.powf(freq_log);
+                let filtconf = config::BiquadParameters::Peaking(config::PeakingWidth::Bandwidth {
+                    freq: freq as PrcFmt,
+                    bandwidth: bw as PrcFmt,
+                    gain: *gain as PrcFmt,
+                });
+                let coeffs = biquad::BiquadCoefficients::from_config(samplerate, filtconf);
+                let filt = biquad::Biquad::new("", samplerate, coeffs);
+                filters.push(filt);
+            }
+        }
+        filters
+    }
+
     pub fn from_config(
         name: &str,
         samplerate: usize,
@@ -210,6 +239,19 @@ impl BiquadCombo {
                     [fls, fp1, fp2, fp3, fhs],
                     [qls, qp1, qp2, qp3, qhs],
                     [gls, gp1, gp2, gp3, ghs],
+                );
+                BiquadCombo {
+                    samplerate,
+                    name,
+                    filters,
+                }
+            }
+            config::BiquadComboParameters::GraphicEqnalizer(params) => {
+                let filters = BiquadCombo::make_graphic(
+                    samplerate,
+                    params.freq_min(),
+                    params.freq_max(),
+                    &params.gains,
                 );
                 BiquadCombo {
                     samplerate,
@@ -309,6 +351,31 @@ pub fn validate_config(samplerate: usize, conf: &config::BiquadComboParameters) 
                 || *fp3 >= maxfreq
             {
                 return Err(config::ConfigError::new("All frequencies must be > 0").into());
+            }
+            Ok(())
+        }
+        config::BiquadComboParameters::GraphicEqnalizer(params) => {
+            if params.freq_min() <= 0.0 || params.freq_max() <= 0.0 {
+                return Err(config::ConfigError::new("Min and max requencies must be > 0").into());
+            } else if params.freq_min() >= maxfreq as f32 || params.freq_max() >= maxfreq as f32 {
+                return Err(config::ConfigError::new(
+                    "Min and max frequencies must be < samplerate/2",
+                )
+                .into());
+            }
+            if params.freq_min() >= params.freq_max() {
+                return Err(config::ConfigError::new(
+                    "Min frequency must be lower than max frequency",
+                )
+                .into());
+            }
+            for gain in params.gains.iter() {
+                if *gain > 20.0 || *gain < -20.0 {
+                    return Err(config::ConfigError::new(
+                        "Equalizer gains must be withing +- 20 dB",
+                    )
+                    .into());
+                }
             }
             Ok(())
         }