+++ CHANGELOG.md
@@ -25,7 +25,8 @@ New features:
   - ShibataHigh441 variant.
   - Shibata and ShibataLow 88.2, 96 and 192.
 - Add GeneralNotch biquad type.
-- Add Tilt equalizer.
+- Add Tilt equalizer biquad combo.
+- Add GraphicEqualizer biquad combo.
 
 Changes:
 - Optimize cpu load of dithering and delay filters.